[Refactor/#168] ViewModelType 분리 -> Domain/Protocols
