+++ UniVoice/UniVoice/Presentation/LoginPage/ViewModel/LoginVM.swift
@@ -8,15 +8,6 @@
 import RxSwift
 import RxCocoa
 
-protocol ViewModelType {
-    associatedtype Input
-    associatedtype Output
-    
-    var disposeBag: DisposeBag { get set }
-    
-    func transform(input: Input) -> Output
-}
-
 enum LoginButtonState {
     case idIsEditingWithoutPW
     case pwIsEditingWithoutID
